requirements.txt
# Computer Vision and Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
scipy>=1.11.0
numpy>=1.24.0

# OCR
pytesseract>=0.3.10

# Google Gemini API
google-generativeai>=0.3.0

# Token counting (for text processing)
tiktoken>=0.5.0

# JSON handling and utilities
python-dotenv>=1.0.0

# Optional: For better image preprocessing
scikit-image>=0.21.0

# PDF Processing
PyMuPDF>=1.23.0
pdf2image>=1.16.0
pypdfium2>=4.0.0

# Web Interface
streamlit>=1.28.0
pandas>=2.0.0